Added SentimentEvaluator Class
I added the Sentiments folder to the project and the AFINN file within
that folder. This is the file that holds all sentiment values. I also
added the SentimentEvaluator class. This class has a constructor which
initializes the term-sentimentValue hashmap, and a method which
tokenizes an input string and returns the entire string's sentiment
value.